% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colsums_with_default}
\alias{colsums_with_default}
\title{Return a non-null default value or compute column sums}
\usage{
colsums_with_default(y, default)
}
\arguments{
\item{y}{SparseMatrix object}

\item{default}{Default column sums}
}
\description{
Return a non-null default value or compute column sums
}
\keyword{internal}
