% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_poisson_residuals.R
\name{compute_pca}
\alias{compute_pca}
\title{Computes PCA for a residual matrix from its cross product}
\usage{
compute_pca(rtr, y, rate, k)
}
\arguments{
\item{rtr}{Cross product of residual matrix.}

\item{y}{Original sparse matrix, scaled by inverse standard deviation.}

\item{rate}{Row-wise rates.}

\item{k}{Number of principal components to return.}
}
\description{
Computes PCA for a residual matrix from its cross product
}
