% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_sparse}
\alias{convert_to_sparse}
\title{Convert a sparse matrix to a SparseMatrix object}
\usage{
convert_to_sparse(y)
}
\arguments{
\item{y}{Sparse matrix (can be a matrix, dgCMatrix, or SparseMatrix).}
}
\description{
Convert a sparse matrix to a SparseMatrix object
}
\keyword{internal}
