% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CountTransform-class}
\alias{CountTransform-class}
\title{Preprocessing on Sparse Count Matrices}
\description{
Preprocessing on Sparse Count Matrices
}
\section{Slots}{

\describe{
\item{\code{func}}{Transformation to be applied to sparse count matrix. Should be a
function with a single parameter.}

\item{\code{center_rows}}{Whether transformed rows should be shifted to have mean
zero.}

\item{\code{center_cols}}{Whether transformed columns should be shifted to have mean
zero.}

\item{\code{scale}}{Whether transformed rows should be scaled to have unit variance.}
}}

