% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{CountTransform}
\alias{CountTransform}
\title{Preprocessing on Sparse Count Matrices}
\usage{
CountTransform(func, center = FALSE, scale = FALSE)
}
\arguments{
\item{func}{Transformation to be applied to sparse count matrix. Should be a
function with a single parameter.}

\item{center}{Whether transformed data should be shifted to have mean zero.
Can be specified either as a vector of two logical values, specifying
whether the rows/columns should be centered, respectively, or a single
logical value specifying whether the columns should be centered (for
consistency with \code{\link[stats:prcomp]{stats::prcomp()}}).}

\item{scale}{Whether transformed rows should be scaled to have unit variance.}
}
\value{
CountTransform object
}
\description{
This class is an abstraction for the preprocessing transformations that are
applied to count data before statistical methods like PCA (e.g., log1p,
scaling, row-centering). Note that these transformations may not preserve the
sparsity of the matrix.
}
\examples{
triple <- CountTransform(function(x) 3 * x, center = FALSE, scale = FALSE)
}
