% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calculateMu}
\alias{.calculateMu}
\title{Calculate mu values from row rates and column sums}
\usage{
.calculateMu(y, nz_ind, rate, n)
}
\arguments{
\item{y}{SparseMatrix object}

\item{rate}{Vector of row rates}

\item{n}{Vector of column sums}
}
\value{
Expected values for non-zero matrix entries
}
\description{
Calculate mu values from row rates and column sums
}
\keyword{internal}
