% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.colsumsWithDefault}
\alias{.colsumsWithDefault}
\title{Return a non-null default value or compute column sums}
\usage{
.colsumsWithDefault(y, default)
}
\arguments{
\item{y}{SparseMatrix object}

\item{default}{Default column sums}
}
\description{
Return a non-null default value or compute column sums
}
\keyword{internal}
