% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.convertToSparse}
\alias{.convertToSparse}
\title{Convert a sparse matrix to a SparseMatrix object}
\usage{
.convertToSparse(y)
}
\arguments{
\item{y}{Sparse matrix (can be a matrix, dgCMatrix, or SparseMatrix).}
}
\description{
Convert a sparse matrix to a SparseMatrix object
}
\keyword{internal}
