% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.nzcols}
\alias{.nzcols}
\title{Get the column indices of non-zero matrix values}
\usage{
.nzcols(y, nz_ind)
}
\arguments{
\item{y}{SparseMatrix object}

\item{nz_ind}{Linear indices of non-zero values}
}
\description{
Get the column indices of non-zero matrix values
}
\keyword{internal}
