% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.nzrows}
\alias{.nzrows}
\title{Get the row indices of non-zero matrix values}
\usage{
.nzrows(y, nz_ind)
}
\arguments{
\item{y}{SparseMatrix object}

\item{nz_ind}{Linear indices of non-zero values}
}
\value{
Row indices of non-zero matrix values
}
\description{
Get the row indices of non-zero matrix values
}
\keyword{internal}
