% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_pca.R
\name{.poissonDevianceResidualsPca}
\alias{.poissonDevianceResidualsPca}
\title{Principal component analysis on deviance residuals}
\usage{
.poissonDevianceResidualsPca(y, k)
}
\arguments{
\item{y}{Sparse matrix (can be a matrix, dgCMatrix, or SparseMatrix).}

\item{k}{Number of principal components to return.}
}
\description{
Principal component analysis on deviance residuals
}
\keyword{internal}
