% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_pca.R
\name{.poissonPearsonResidualsPca}
\alias{.poissonPearsonResidualsPca}
\title{Principal component analysis on Pearson residuals}
\usage{
.poissonPearsonResidualsPca(y, k)
}
\arguments{
\item{y}{Sparse matrix (can be a matrix, dgCMatrix, or SparseMatrix)}

\item{k}{Number of principal components to return}
}
\value{
List with components:
\itemize{
  \item{sdev}{Standard deviations of principal components}
  \item{rotation}{Matrix of variable loadings (i.e., matrix containing the
  eigenvectors of the covariance/correlation matrix as columns)}
  \item{x}{Matrix of rotated data (rotated after applying the transformations
  specified)}
}
}
\description{
Principal component analysis on Pearson residuals
}
\keyword{internal}
