% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_pca.R
\name{.rawResidualsPca}
\alias{.rawResidualsPca}
\title{Principal component analysis on raw residuals}
\usage{
.rawResidualsPca(y, k, row_offset, col_offset)
}
\arguments{
\item{y}{Sparse matrix (can be a matrix, dgCMatrix, or SparseMatrix).}

\item{k}{Number of principal components to return.}

\item{row_offset, col_offset}{Vectors whose product subtracted from \code{y} gives
the residual matrix.}
}
\description{
Principal component analysis on raw residuals
}
\keyword{internal}
