% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rowRatesWithDefault}
\alias{.rowRatesWithDefault}
\title{Return a non-null default value or compute row-wise rates}
\usage{
.rowRatesWithDefault(y, default)
}
\arguments{
\item{y}{SparseMatrix object}

\item{default}{Default row-wise rates}
}
\description{
Return a non-null default value or compute row-wise rates
}
\keyword{internal}
