% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safeDivide}
\alias{.safeDivide}
\title{Divide a by b, returning zero if b is zero}
\usage{
.safeDivide(a, b)
}
\arguments{
\item{a}{Numerator}

\item{b}{Denominator}
}
\description{
Divide a by b, returning zero if b is zero
}
\keyword{internal}
