% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparse_matrix.R
\name{get_decompressed_matrix_file}
\alias{get_decompressed_matrix_file}
\title{Unzip (and untar) a .(t)gz or .(t)bz2 sparse matrix file}
\usage{
get_decompressed_matrix_file(file)
}
\arguments{
\item{file}{character(1) path to a potentially compressed sparse matrix file.}
}
\value{
character(1) path to unzipped file contents.
}
\description{
Unzip (and untar) a .(t)gz or .(t)bz2 sparse matrix file
}
\note{
This function is a no-op for all other file extensions.
}
