% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nz_rows}
\alias{get_nz_rows}
\title{Get the row indices of non-zero matrix values}
\usage{
get_nz_rows(y, nz_ind)
}
\arguments{
\item{y}{SparseMatrix object}

\item{nz_ind}{Linear indices of non-zero values}
}
\description{
Get the row indices of non-zero matrix values
}
\keyword{internal}
