% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_rates.R
\name{group_rates}
\alias{group_rates}
\title{Row-wise Rates for Groups}
\usage{
group_rates(y, g)
}
\arguments{
\item{y}{Sparse matrix (can be a base matrix, dgCMatrix, or SparseMatrix).}

\item{g}{Factor specifying the group for each column.}
}
\description{
Row-wise Rates for Groups
}
