% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_rates.R
\name{groupRates}
\alias{groupRates}
\title{Row-wise Rates for Groups}
\usage{
groupRates(y, g)
}
\arguments{
\item{y}{Sparse matrix (can be a matrix, dgCMatrix, or SparseMatrix).}

\item{g}{Factor specifying the group for each column.}
}
\description{
Row-wise Rates for Groups
}
