% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{identity_transform}
\alias{identity_transform}
\title{Identity Transformation}
\usage{
identity_transform(center = FALSE, scale = FALSE)
}
\arguments{
\item{center}{Whether transformed data should be shifted to have mean zero.
Can be specified either as a vector of two logical values, specifying
whether the rows/columns should be centered, respectively, or a single
logical value specifying whether the columns should be centered (for
consistency with \code{\link{stats}{prcomp()}}).}

\item{scale}{Whether transformed rows should be scaled to have unit variance.}
}
\value{
CountTransform object.
}
\description{
CountTransform mapping x to itself.
}
\note{
Internally, the identity function is represented by a NULL value.
}
