% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxx.R
\name{parse_sparse_csv}
\alias{parse_sparse_csv}
\title{use RcppArmadillo to parse 'sparse csv'}
\usage{
parse_sparse_csv(path)
}
\arguments{
\item{path}{character(1) path to uncompressed CSV with no header}
}
\value{
list with component sp, a dgCMatrix instance
}
\description{
use RcppArmadillo to parse 'sparse csv'
}
\examples{
# superficial
if (!requireNamespace("bench")) stop("install bench package to run this example")
pa = system.file("extdata/tenx_subset-transpose.csv.gz", package="smallcount")
tf = tempfile()
file.copy(pa, tf)
file.rename(tf, ntf <- paste0(tf, ".gz"))
system(paste("gunzip", ntf)) # restores tf as gunzipped pa
tf2 = tempfile()
system(sprintf("sed -e '1,1d' \%s | cut -d ',' -f 2- > \%s", tf, tf2))
tim = bench::mark(mm <- parse_sparse_csv(tf2), iterations=2)
tim
dim(mm$sp)
mm$sp[11:16,1:6]
}
