% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_poisson_residuals.R
\name{pca_poisson_raw_residuals}
\alias{pca_poisson_raw_residuals}
\title{Principal component analysis on raw residuals}
\usage{
pca_poisson_raw_residuals(y, k, rate, n)
}
\arguments{
\item{y}{Sparse matrix (can be a base matrix, dgCMatrix, or SparseMatrix).}

\item{k}{Number of principal components to return.}

\item{rate}{Row-wise rates.}

\item{n}{Total counts in each column.}
}
\description{
Principal component analysis on raw residuals
}
