% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_poisson_residuals.R
\name{pca_poisson_residuals}
\alias{pca_poisson_residuals}
\title{Principal component analysis on residuals}
\usage{
pca_poisson_residuals(y, k = 50, rate = NULL, n = NULL, residual = "raw")
}
\arguments{
\item{y}{Sparse matrix (can be a base matrix, dgCMatrix, or SparseMatrix).}

\item{k}{Number of principal components to return. Default 50.}

\item{rate}{Row-wise rates.}

\item{n}{Total counts in each column.}

\item{residual}{Residual type ("raw" or "pearson"). Default raw.}
}
\description{
Principal component analysis on residuals
}
\examples{
data(tenx_subset)
dim(tenx_subset)
system.time({
  pc <- pca_poisson_residuals(tenx_subset, k = 10, residual = "pearson")
})
plot(pc$x[,1:2])
barplot(pc$sdev)
}
