% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_poisson_residuals.R
\name{pca_poisson_residuals_new}
\alias{pca_poisson_residuals_new}
\title{Principal Component Analysis on Pearson residuals}
\usage{
pca_poisson_residuals_new(y, k = 50, rate = NULL, n = NULL)
}
\arguments{
\item{y}{SparseMatrix object.}

\item{k}{Number of principal components to return.}

\item{rate}{Row-wise rates.}

\item{n}{Total counts in each column.}
}
\description{
Principal Component Analysis on Pearson residuals
}
