% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_deviance.R
\name{poisson_deviance}
\alias{poisson_deviance}
\title{Poisson Deviance}
\usage{
poisson_deviance(y, rate = NULL, n = NULL)
}
\arguments{
\item{y}{Sparse matrix (can be a base matrix, dgCMatrix, or SparseMatrix).}

\item{rate}{Row-wise rates.}

\item{n}{Total counts in each column.}
}
\description{
Poisson Deviance
}
\examples{
data("tenx_subset")
dev <- poisson_deviance(tenx_subset)
hist(dev, nclass = 50)
}
