% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_deviance.R
\name{poisson_deviance_new}
\alias{poisson_deviance_new}
\title{Poisson Deviance}
\usage{
poisson_deviance_new(y, rate = NULL, n = NULL)
}
\arguments{
\item{y}{Sparse matrix (can be a base matrix, dgCMatrix, or SparseMatrix).}

\item{rate}{Row-wise rates.}

\item{n}{Total counts in each column.}
}
\description{
Poisson Deviance
}
