% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dispersion.R
\name{poisson_dispersion}
\alias{poisson_dispersion}
\title{Compute Poisson Dispersion}
\usage{
poisson_dispersion(y, rate = NULL, n = NULL)
}
\arguments{
\item{y}{A tgCMatrix sparse Matrix.}

\item{rate}{The rowwise rates.}

\item{n}{The total counts in each column.}
}
\description{
Compute Poisson Dispersion
}
