% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dispersion.R
\name{poisson_dispersion_new}
\alias{poisson_dispersion_new}
\title{Compute Poisson Dispersion}
\usage{
poisson_dispersion_new(y, rate = NULL, n = NULL)
}
\arguments{
\item{y}{Sparse matrix (can be a base matrix, dgCMatrix, or SparseMatrix).}

\item{rate}{Row-wise rates.}

\item{n}{Total counts in each column.}
}
\description{
Compute Poisson Dispersion
}
