% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_pca.R
\name{poisson_pca}
\alias{poisson_pca}
\title{Principal component analysis on Poisson data}
\usage{
poisson_pca(y, k = 50, transform = NULL, center = FALSE, scale = FALSE)
}
\arguments{
\item{y}{Sparse matrix (can be a base matrix, dgCMatrix, or SparseMatrix).}

\item{k}{Number of principal components to return. Default 50.}

\item{transform}{CountTransform object or character(1) specifying a
transformation to apply to \code{y}. Arguments \code{center} and \code{scale} are only
applied when \code{transform} is a character string. Accepted strings are:
\code{"id"} (or \code{NULL}) for the trivial identity function, \code{"log1p"} for
\code{log(x + 1)}, \code{"cpm_log1p"} for \code{log(x/1e6 + 1)}, and \code{"med_log1p"} for
\code{log(x/med + 1)}, where \code{med = median(colSums(y))}.}

\item{center}{Whether transformed data should be shifted to have mean zero.
Can be specified either as a vector of two logical values, specifying
whether the rows/columns should be centered, respectively, or a single
logical value specifying whether the columns should be centered (for
consistency with \code{\link[stats:prcomp]{stats::prcomp()}}).}

\item{scale}{Whether transformed rows should be scaled to have unit variance.}
}
\description{
Principal component analysis on Poisson data
}
\examples{
data(tenx_subset)
dim(tenx_subset)
system.time({
  pc <- poisson_pca(tenx_subset, k = 10, transform = "pearson")
})
plot(pc$x[,1:2])
barplot(pc$sdev)
}
