% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_deviance.R
\name{poissonDeviance}
\alias{poissonDeviance}
\title{Poisson Deviance}
\usage{
poissonDeviance(y, rate = NULL, n = NULL)
}
\arguments{
\item{y}{Sparse matrix (can be a matrix, dgCMatrix, or SparseMatrix)}

\item{rate}{Row-wise rates}

\item{n}{Total counts in each column}
}
\value{
Row-wise Poisson deviance
}
\description{
Poisson Deviance
}
\examples{
data("tenx_subset")
dev <- poissonDeviance(tenx_subset)
hist(dev, nclass = 50)
}
