% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dispersion.R
\name{poissonDispersion}
\alias{poissonDispersion}
\title{Poisson Dispersion}
\usage{
poissonDispersion(y, rate = NULL, n = NULL)
}
\arguments{
\item{y}{Sparse matrix (can be a matrix, dgCMatrix, or SparseMatrix).}

\item{rate}{Row-wise rates.}

\item{n}{Total counts in each column.}
}
\value{
Row-wise Poisson dispersion
}
\description{
Poisson Dispersion
}
\examples{
data("tenx_subset")
disp <- poissonDispersion(tenx_subset)
hist(disp, nclass = 50)
}
