% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparse_csv.R
\name{read_sparse_csv_transpose}
\alias{read_sparse_csv_transpose}
\title{Read genes in columns}
\usage{
read_sparse_csv_transpose(file, sep = ",")
}
\arguments{
\item{file}{character(1) path to a gzipped delimited file with counts; features are rows, cells are columns}

\item{sep}{character(1) delimiter for fields (defaults to ',')}
}
\description{
Read genes in columns
}
