% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparse_matrix.R
\name{read_sparse_matrix}
\alias{read_sparse_matrix}
\title{Read a SparseMatrix object from a file}
\usage{
read_sparse_matrix(file, representation = "svt")
}
\arguments{
\item{file}{character(1) path to a sparse matrix file (.h5, .mtx, and .csv
extensions are supported, compressed or decompressed).}

\item{representation}{character(1) internal SparseMatrix representation,
either "svt" or "coo". (Default "svt").}
}
\value{
SparseMatrix object.
}
\description{
Read a SparseMatrix object from a file
}
