% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparse_matrix.R
\name{readSparseMatrix}
\alias{readSparseMatrix}
\title{Load data from a 10X Genomics experiment}
\usage{
readSparseMatrix(
  sample,
  col.names = FALSE,
  row.names = c("id", "symbol"),
  genome = NULL
)
}
\arguments{
\item{sample}{character(1) directory name corresponding to a 10X sample. The
directory should contain a matrix file, a gene/feature annotation file, and
a barcode annotation file.

Alternatively, the string may contain a path to a HDF5 file in the sparse
matrix format generated by 10X.

Alternatively, the string may contain a prefix of names for the three-file
system described above, where the rest of the name of each file follows the
standard 10X output.}

\item{col.names}{logical(1) indicating whether the columns of the matrix
should be named with the cell barcodes.}

\item{row.names}{character(1) specifying whether to use Ensembl IDs ("id") or
gene symbols ("symbol") as row names. If using symbols, the Ensembl ID will
be appended to disambiguate in case the same symbol corresponds to multiple
Ensembl IDs.}

\item{genome}{character(1) specifying the genome for HDF5 files output by
CellRanger v2.}

\item{representation}{character(1) specifying the internal SparseMatrix
representation (Default "svt").}
}
\value{
A \linkS4class{SparseMatrix} object containing count data for each
gene (row) and cell (column) in \code{sample}.
}
\description{
Creates a \linkS4class{SparseMatrix} from the CellRanger output directories
for 10X Genomics data.
}
\details{
The signature of this function and its corresponding documentation
has largely been adapted from the \code{Read10xCounts} function in the
\pkg{DropletUtils} package.

Note that user-level manipulation of sparse matrices requires loading of the
\pkg{SparseArray} package. Otherwise, calculation of \code{rowSums},
\code{colSums}, etc. will result in errors.
}
\examples{
data("tenx_subset")  # Original dataset
new <- readSparseMatrix(system.file(
  "extdata/tenx_subset.csv.gz", package = "smallcount"))
identical(new, tenx_subset)

}
\references{
Zheng GX, Terry JM, Belgrader P, and others (2017). Massively
parallel digital transcriptional profiling of single cells. \emph{Nat Commun}
8:14049.

10X Genomics (2017). Gene-Barcode Matrices.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/2.2/output/matrices}

10X Genomics (2018). Feature-Barcode Matrices.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/matrices}

10X Genomics (2018). HDF5 Gene-Barcode Matrix Format.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/2.2/advanced/h5_matrices}

10X Genomics (2018). HDF5 Feature Barcode Matrix Format.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/advanced/h5_matrices}
}
