% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_rates_with_default}
\alias{row_rates_with_default}
\title{Return a non-null default value or compute row-wise rates}
\usage{
row_rates_with_default(y, default)
}
\arguments{
\item{y}{SparseMatrix object}

\item{default}{Default row-wise rates}
}
\description{
Return a non-null default value or compute row-wise rates
}
\keyword{internal}
