% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_divide}
\alias{safe_divide}
\title{Divide a by b, returning zero if b is zero}
\usage{
safe_divide(a, b)
}
\arguments{
\item{a}{Numerator}

\item{b}{Denominator}
}
\description{
Divide a by b, returning zero if b is zero
}
\keyword{internal}
