% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{scaled_log1p_transform}
\alias{scaled_log1p_transform}
\title{Scaled Log1p Transformation}
\usage{
scaled_log1p_transform(coef, center = FALSE, scale = FALSE)
}
\arguments{
\item{coef}{Scaling coefficient.}

\item{center}{Whether transformed data should be shifted to have mean zero.
Can be specified either as a vector of two logical values, specifying
whether the rows/columns should be centered, respectively, or a single
logical value specifying whether the columns should be centered (for
consistency with \code{\link[stats:prcomp]{stats::prcomp()}}).}

\item{scale}{Whether transformed rows should be scaled to have unit variance.}
}
\value{
CountTransform object.
}
\description{
Maps x to log(\code{coef} * x + 1).
}
