% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenx_subset}
\alias{tenx_subset}
\title{compressed csv file with TENx brain data subset, genes in rows}
\format{
dgCMatrix
}
\usage{
tenx_subset
}
\description{
compressed csv file with TENx brain data subset, genes in rows
}
\keyword{datasets}
