% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenx_subset_new}
\alias{tenx_subset_new}
\title{Compressed csv file with TENx brain data subset, genes in rows}
\format{
SparseMatrix
}
\usage{
tenx_subset_new
}
\description{
Compressed csv file with TENx brain data subset, genes in rows
}
\keyword{datasets}
