% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TransformedMatrix-class}
\alias{TransformedMatrix-class}
\title{Transformed Count Matrix}
\description{
Transformed Count Matrix
}
\section{Slots}{

\describe{
\item{\code{y}}{SparseMatrix object.}

\item{\code{row_offset,col_offset}}{Vectors whose product
`outer(row_offset, col_offset)` represents the residual between `y` and a
dense transformation of `y` (e.g., row-centered `y`).}
}}

