% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparse_matrix.R
\name{unzip_file}
\alias{unzip_file}
\title{Decompress (and untar) .(t)gz or .(t)bz2 sparse matrix data at a file path}
\usage{
unzip_file(filepath, temp_dir = tempdir())
}
\arguments{
\item{filepath}{character(1) path to potentially compressed matrix data.}

\item{temp_dir}{character(1) temporary directory for unzipped files.}
}
\value{
character(1) path to unzipped file contents.
}
\description{
Decompress (and untar) .(t)gz or .(t)bz2 sparse matrix data at a file path
}
\note{
This function is a no-op for all other file extensions.
}
\keyword{internal}
