% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparse_matrix.R
\name{validate_sample}
\alias{validate_sample}
\title{Check that a path is a valid file, file directory, or file prefix}
\usage{
validate_sample(filepath)
}
\arguments{
\item{filepath}{character(1) path to matrix data.}
}
\value{
character(1) potentially updated file path.
}
\description{
Check that a path is a valid file, file directory, or file prefix
}
\keyword{internal}
