% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_pca.R
\name{compute_pca}
\alias{compute_pca}
\title{Compute the eigendecomposition of \code{R \%*\% t(R)} for a residual matrix \code{R}}
\usage{
compute_pca(rtr, k, y, offset1 = NULL, offset2 = NULL)
}
\arguments{
\item{rtr}{Cross product of residual matrix \code{R = y - offset1 \%*\% offset2}.}

\item{k}{Number of principal components to return.}

\item{y}{Sparse matrix.}

\item{offset1, offset2}{Vectors whose product is the difference between \code{y}
and the residual matrix.}
}
\description{
Compute the eigendecomposition of \code{R \%*\% t(R)} for a residual matrix \code{R}
}
\keyword{internal}
