% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{CountTransform}
\alias{CountTransform}
\title{CountTransform Constructor}
\usage{
CountTransform(func, center = FALSE, scale = FALSE)
}
\arguments{
\item{func}{Transformation to be applied to sparse count matrix. Should be a
function with a single parameter.}

\item{center}{Whether transformed data should be shifted to have mean zero.
Can be specified either as a vector of two logical values, specifying
whether the rows/columns should be centered, respectively, or a single
logical value specifying whether the columns should be centered (for
consistency with \code{\link[stats]{prcomp()}}).}

\item{scale}{Whether transformed rows should be scaled to have unit variance}
}
\value{
CountTransform object
}
\description{
CountTransform Constructor
}
\examples{
triple <- CountTransform(function(x) 3 * x, center = FALSE, scale = FALSE)
}
