% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_rates.R
\name{group_rates}
\alias{group_rates}
\title{Rowwise rates for groups}
\usage{
group_rates(y, g)
}
\arguments{
\item{y}{A tgCMatrix sparse Matrix.}

\item{g}{A factor defining the group for each column.}
}
\description{
Rowwise rates for groups
}
