% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxx.R
\name{parse_sparse_csv}
\alias{parse_sparse_csv}
\title{use RcppArmadillo to parse 'sparse csv'}
\usage{
parse_sparse_csv(fname)
}
\arguments{
\item{fname}{character(1) path to simple sparse csv no header, no rownames}
}
\description{
use RcppArmadillo to parse 'sparse csv'
}
