% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_poisson_residuals.R
\name{pca_poisson_residuals}
\alias{pca_poisson_residuals}
\title{Principal Component Analysis on Pearson residuals}
\usage{
pca_poisson_residuals(y, k = 50, rate = NULL, n = NULL)
}
\arguments{
\item{y}{A tgCMatrix sparse Matrix.}

\item{k}{Number of principal components to return.}

\item{rate}{The rowwise rates.}

\item{n}{The total counts in each column.}
}
\description{
Principal Component Analysis on Pearson residuals
}
