% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_dispersion.R
\name{poisson_dispersion}
\alias{poisson_dispersion}
\title{Poisson Dispersion}
\usage{
poisson_dispersion(y, rate = NULL, n = NULL)
}
\arguments{
\item{y}{Sparse matrix (can be a base matrix, dgCMatrix, or SparseMatrix).}

\item{rate}{Row-wise rates.}

\item{n}{Total counts in each column.}
}
\description{
Poisson Dispersion
}
\examples{
data("tenx_subset")
disp <- poisson_dispersion(tenx_subset)
hist(disp, nclass = 50)
}
