% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenx_subset}
\alias{tenx_subset}
\title{Compressed TENx brain data subset, genes in rows}
\format{
SparseMatrix
}
\usage{
tenx_subset
}
\description{
Compressed TENx brain data subset, genes in rows
}
\keyword{datasets}
