% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TransformedMatrix-class}
\alias{TransformedMatrix-class}
\title{Transformed Count Matrix}
\description{
Representation of a sparse count matrix after a CountTransform is applied.
}
\section{Slots}{

\describe{
\item{\code{y}}{SparseMatrix object.}

\item{\code{row_offset,col_offset}}{Vectors whose product
\code{outer(row_offset, col_offset)} represents the residual between \code{y} and a
dense transformation of \code{y} (e.g., row-centered \code{y}).}
}}

