% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{TransformedMatrix}
\alias{TransformedMatrix}
\title{Transformed Count Matrix}
\usage{
TransformedMatrix(y, transform)
}
\arguments{
\item{y}{SparseMatrix object.}

\item{transform}{Transformation to apply to \code{y}.}
}
\value{
TransformedMatrix object.
}
\description{
Transformed Count Matrix
}
\examples{
mat <- as(matrix(c(1:9), nrow=3, ncol=3), "SVT_SparseMatrix")
triple <- CountTransform(function(x) 3 * x, center = FALSE, scale = FALSE)
tripled_mat <- TransformedMatrix(mat, triple)
}
